Sig-scalability test segregation to allow for assignment and cleaning, and easier refactoring
