+++ test/e2e/BUILD
@@ -26,7 +26,6 @@ go_library(
         "daemon_restart.go",
         "daemon_set.go",
         "dashboard.go",
-        "density.go",
         "deployment.go",
         "disruption.go",
         "dns.go",
@@ -39,7 +38,6 @@ go_library(
         "etcd_failure.go",
         "events.go",
         "example_cluster_dns.go",
-        "example_k8petstore.go",
         "examples.go",
         "firewall.go",
         "garbage_collector.go",
@@ -57,7 +55,6 @@ go_library(
         "kubelet.go",
         "kubelet_perf.go",
         "limit_range.go",
-        "load.go",
         "logging_soak.go",
         "monitoring.go",
         "namespace.go",
@@ -151,6 +148,7 @@ go_library(
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
         "//test/e2e/generated:go_default_library",
+        "//test/e2e/perf:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/upgrades:go_default_library",
         "//test/e2e_federation:go_default_library",
@@ -202,10 +200,8 @@ go_library(
         "//vendor:k8s.io/client-go/pkg/apis/policy/v1beta1",
         "//vendor:k8s.io/client-go/rest",
         "//vendor:k8s.io/client-go/tools/cache",
-        "//vendor:k8s.io/client-go/transport",
         "//vendor:k8s.io/client-go/util/flowcontrol",
         "//vendor:k8s.io/client-go/util/integer",
-        "//vendor:k8s.io/client-go/util/workqueue",
     ],
 )
 
@@ -226,6 +222,7 @@ go_test(
         "//pkg/client/clientset_generated/clientset:go_default_library",
         "//pkg/metrics:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/perf:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/utils:go_default_library",
         "//vendor:github.com/onsi/ginkgo",
@@ -263,6 +260,7 @@ filegroup(
         "//test/e2e/common:all-srcs",
         "//test/e2e/framework:all-srcs",
         "//test/e2e/generated:all-srcs",
+        "//test/e2e/perf:all-srcs",
         "//test/e2e/perftype:all-srcs",
         "//test/e2e/scheduling:all-srcs",
         "//test/e2e/testing-manifests:all-srcs",
