+++ test/e2e/empty.go
@@ -22,6 +22,7 @@ import (
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/v1"
 	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/kubernetes/test/e2e/perf"
 
 	. "github.com/onsi/ginkgo"
 )
@@ -42,7 +43,7 @@ var _ = framework.KubeDescribe("Empty [Feature:Empty]", func() {
 	})
 
 	It("starts a pod", func() {
-		configs, _ := generateConfigsForGroup([]*v1.Namespace{f.Namespace}, "empty-pod", 1, 1, framework.GetPauseImageName(f.ClientSet), []string{}, api.Kind("ReplicationController"), 0)
+		configs, _ := perf.GenerateConfigsForGroup([]*v1.Namespace{f.Namespace}, "empty-pod", 1, 1, framework.GetPauseImageName(f.ClientSet), []string{}, api.Kind("ReplicationController"), 0)
 		if len(configs) != 1 {
 			framework.Failf("generateConfigs should have generated single config")
 		}
