+++ test/e2e/perf/OWNERS
@@ -1,0 +1,7 @@
+approvers:
+- gmarek
+- wojtek-t
+reviewers:
+- gmarek
+- wojtek-t
+- timothysc
